+++ Changelog.md
@@ -4,11 +4,12 @@
 
 ### Features and Improvements
 
-- None
+- Add webhook and pause schedule actions to automations - [#818](https://github.com/PrefectHQ/ui/pull/818)
 
 ### Bugfixes
 
 - Display task run name in Schematic - [#863](https://github.com/PrefectHQ/ui/pull/863)
+- Various bugfixes related to null tenants and team invitaitons - [#862](https://github.com/PrefectHQ/ui/pull/862)
 
 ## 2021-05-25